requirements.txt
requests>=2.31.0
geopy>=2.4.1
pandas>=2.2.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
tabulate>=0.9.0
Pillow>=10.2.0
Flask>=3.0.0
Flask-CORS>=4.0.0
Werkzeug>=3.0.0
python-docx>=1.1.0
